delete cstudent

+++ src/services/studentClassService.js
@@ -1,3 +1,4 @@
+import { Op } from "sequelize";
 import ErrorMessage from "../common/errorMessage";
 import HttpException from "../errors/httpException";
 import db from "../models";
@@ -52,22 +53,15 @@ export const deleteStudentInClass = async (studentId, classId) => {
   }
   if (!studentId) throw new HttpException(422, ErrorMessage.MISSING_PARAMETER);
   const classRoom = await classService.getClassById(classId);
-
+  console.log(new Date(classRoom.timeStart) <= new Date());
+  if (new Date(classRoom.timeStart) <= new Date()) {
+    throw new HttpException(
+      422,
+      ErrorMessage.CUSTOM("Class has started, can't removed students")
+    );
+  }
   const data = await db.StudentClass.destroy({
     where: { classId: classId, studentId: studentId },
-    include: [
-      {
-        model: db.Class,
-        where: {
-          timeStart: {
-            [Op.lt]: new Date(), // Lấy những lớp học có thời gian bắt đầu nhỏ hơn ngày hiện tại
-          },
-        },
-        required: true, // Sử dụng `required: true` để thực hiện INNER JOIN thay vì LEFT JOIN
-        // nest: true,
-        // raw: false,
-      },
-    ],
   });
 
   return data;
